+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [1.7.2]
+### Fixed
+- [Core] Add a function to close the `<Modal>` when users click `<Overlay>` and the onClose prop exists. (#140)
+
+### Changed
+- [Core] Update styles for `<ListRow>` to match design in nested lists. (#141)
+- [Core] Update `<Button>` to support rendering into custom HTML tags. (#144)
+- [Storybook] Add the showcase of two overlaying `<Modal>`. (#140)
+- [Storybook] Add examples for nested `<List>`. (#141)
+
 ## [1.7.1]
 ### Fixed
 - [Core] Fix `<Modal>` not rendering its content (#139)